feat(ui): cleanup and refactor
