+++ src/pages/bridge/faucet/index.tsx
@@ -76,7 +76,7 @@ export default function Faucet(): JSX.Element {
       /> */}
 
       <Container maxWidth="2xl" className="space-y-6">
-        <DoubleGlowShadow /*opacity="0.6"*/>
+        <DoubleGlowShadow>
           <div className="p-4 space-y-4 rounded bg-dark-900" style={{ zIndex: 1 }}>
             <div className="flex items-center justify-center mb-4 space-x-3">
               <div className="grid grid-cols-3 rounded p-3px bg-dark-800 h-[46px]">
