+++ src/pages/bridge/index.tsx
@@ -1,58 +1,30 @@
 import {
-  AbstractCurrency,
   ChainId,
   Currency,
-  CurrencyAmount,
-  JSBI,
-  SmartBCH,
-  NATIVE,
   Token,
-  WNATIVE,
 } from '@mistswapdex/sdk'
 import React, { useCallback, useEffect, useState } from 'react'
-let json;
 import { AutoRow } from '../../components/Row'
 import Container from '../../components/Container'
 import Head from 'next/head'
-import { ArrowDown, ArrowRight, Clock, Settings } from 'react-feather'
+import { ArrowRight } from 'react-feather'
 import Typography from '../../components/Typography'
 import Web3Connect from '../../components/Web3Connect'
 import { t } from '@lingui/macro'
 import { useActiveWeb3React } from '../../hooks/useActiveWeb3React'
 import { useLingui } from '@lingui/react'
-import { /*useMultichainCurrencyBalance,*/ useTokenBalance } from '../../state/wallet/hooks'
+import { useETHBalances } from '../../state/wallet/hooks'
 import DoubleGlowShadow from '../../components/DoubleGlowShadow'
-import SolarbeamLogo from '../../components/SolarbeamLogo'
 import { BottomGrouping } from '../../features/exchange-v1/swap/styleds'
 import Button from '../../components/Button'
 import DualChainCurrencyInputPanel from '../../components/DualChainCurrencyInputPanel'
 import ChainSelect from '../../components/ChainSelect'
 import { QuestionMarkCircleIcon } from '@heroicons/react/solid'
-
-export type Chain = {
-  id: ChainId
-  name?: string
-  icon?: string
-}
-
-import { useBridgeInfo } from '../../features/bridge/hooks'
-import useSWR, { SWRResponse } from 'swr'
-import { getAddress } from '@ethersproject/address'
-import { classNames, formatNumber, formatPrice } from '../../functions'
-import { SUPPORTED_NETWORKS } from '../../modals/ChainModal'
-// import { NETWORK_ICON, NETWORK_LABEL } from '../../constants/networks'
+import { classNames, formatNumber } from '../../functions'
 import { ethers } from 'ethers'
-// import { useAnyswapTokenContract, useTokenContract } from '../../hooks'
-import Loader from '../../components/Loader'
-import { getWeb3ReactContext, useWeb3React } from '@web3-react/core'
+import { useWeb3React } from '@web3-react/core'
 import NavLink from '../../components/NavLink'
-import { useRouter } from 'next/router'
-import Modal from '../../components/Modal'
-import ModalHeader from '../../components/ModalHeader'
-import { NETWORK_ICON, NETWORK_LABEL } from '../../config/networks'
-import { useTokenContract } from '../../hooks'
 
-import { default as bridge } from './bridge.json';
 import { useEthPrice } from '../../services/graph';
 import BridgeModal from '../../modals/BridgeModal';
 import { xaiGetPermissions, xaiQuote } from '../../services/sideshift.ai';
@@ -60,184 +32,7 @@ import { getBchPoolBalance, getSmartBchPoolBalance, HopDirection, randomId } fro
 import { useTransactionUpdater } from '../../state/bridgeTransactions/hooks';
 import { TransactionDetails } from '../../state/bridgeTransactions/reducer';
 import CashAddressInput from '../../components/Input/Cashaddress';
-import QuestionHelper from '../../components/QuestionHelper';
-
-type BridgeDataInfo = {
-  methodId: string,
-  symbol: string,
-  hint: string,
-  name: string,
-  logoUrl: string,
-  title: string,
-  chainId: number,
-}
-
-type AnyswapTokenInfo = {
-  ID: string
-  Name: string
-  Symbol: string
-  Decimals: number
-  Description: string
-  BaseFeePercent: number
-  BigValueThreshold: number
-  DepositAddress: string
-  ContractAddress: string
-  DcrmAddress: string
-  DisableSwap: boolean
-  IsDelegateContract: boolean
-  MaximumSwap: number
-  MaximumSwapFee: number
-  MinimumSwap: number
-  MinimumSwapFee: number
-  PlusGasPricePercentage: number
-  SwapFeeRate: number
-  FeeUsd?: number
-  SwapRate?: number
-}
-
-type AnyswapResultPairInfo = {
-  DestToken: AnyswapTokenInfo
-  PairID: string
-  SrcToken: AnyswapTokenInfo
-  destChainID: string
-  logoUrl: string
-  name: string
-  srcChainID: string
-  symbol: string
-}
-
-type AvailableChainsInfo = {
-  id: string
-  token: AnyswapTokenInfo
-  other: AnyswapTokenInfo
-  logoUrl: string
-  name: string
-  symbol: string
-  destChainID: string
-}
-
-export type AnyswapTokensMap = { [chainId: number]: { [contract: string]: AvailableChainsInfo } }
-
-const bridgeData: BridgeDataInfo[] = bridge;
-const ourTokenInfo = { Symbol: "BCH", Name: "Bitcoin Cash", Decimals: 18, ContractAddress: "bch" } as AnyswapTokenInfo
-
-const anyswapInfo: AnyswapTokensMap = (() => {
-  const data = {}
-  bridgeData.forEach(val => data[val.methodId] = val)
-
-  let result: AnyswapTokensMap = {}
-
-  Object.keys(data || {}).map((key) => {
-    const bridgeInfo: BridgeDataInfo = data[key]
-    const info: AnyswapResultPairInfo = [bridgeInfo].map(val => {
-      return {
-        SrcToken: {
-          Decimals: 18,
-          Symbol: val.symbol,
-          ID: val.methodId,
-          Name: val.title,
-          ContractAddress: val.methodId
-        } as AnyswapTokenInfo,
-        DestToken: ourTokenInfo,
-        PairID: val.methodId,
-        destChainID: String(ChainId.SMARTBCH),
-        srcChainID: String(val.chainId),
-        logoUrl: val.logoUrl,
-        name: val.title,
-        symbol: val.symbol
-      } as AnyswapResultPairInfo
-    })[0]
-
-    let sourceContractAddress = info.SrcToken.ContractAddress
-    // if (!sourceContractAddress) {
-    //   sourceContractAddress = WNATIVE[parseInt(info.srcChainID)].address
-    // }
-
-    // sourceContractAddress = sourceContractAddress.toLowerCase()
-
-    let existingSource = result[parseInt(info.srcChainID)]
-    if (!existingSource) {
-      result[parseInt(info.srcChainID)] = {
-        [sourceContractAddress]: {
-          destChainID: info.destChainID,
-          id: info.PairID,
-          logoUrl: info.logoUrl,
-          name: info.name,
-          symbol: info.symbol,
-          token: info.DestToken,
-          other: info.SrcToken,
-        },
-      }
-    } else {
-      result[parseInt(info.srcChainID)][sourceContractAddress] = {
-        destChainID: info.destChainID,
-        id: info.PairID,
-        logoUrl: info.logoUrl,
-        name: info.name,
-        symbol: info.symbol,
-        token: info.DestToken,
-        other: info.SrcToken,
-      }
-    }
-
-    let destContractAddress = info.DestToken.ContractAddress
-    // if (!destContractAddress) {
-    //   destContractAddress = WNATIVE[parseInt(info.destChainID)].address
-    // }
-
-    // destContractAddress = destContractAddress.toLowerCase()
-
-    let existingDestination = result[parseInt(info.destChainID)]
-    if (!existingDestination) {
-      result[parseInt(info.destChainID)] = {
-        [destContractAddress]: {
-          destChainID: info.srcChainID,
-          id: info.PairID,
-          logoUrl: info.logoUrl,
-          name: info.name,
-          symbol: info.symbol,
-          token: info.SrcToken,
-          other: info.DestToken,
-        },
-      }
-    } else {
-      result[parseInt(info.destChainID)][destContractAddress] = {
-        destChainID: info.srcChainID,
-        id: info.PairID,
-        logoUrl: info.logoUrl,
-        name: info.name,
-        symbol: info.symbol,
-        token: info.SrcToken,
-        other: info.DestToken,
-      }
-    }
-  })
-
-  const nativeBch = result[0]["bch"]
-  result[ChainId.SMARTBCH] = {
-    "bch": {
-        destChainID: "0",
-        id: nativeBch.id,
-        logoUrl: nativeBch.logoUrl,
-        name: nativeBch.name,
-        symbol: nativeBch.symbol,
-        token: ourTokenInfo,
-        other: nativeBch.token,
-    }
-  }
-
-  return result
-})()
-
-const chains: { [chainId: number]: Chain } = {}
-const chainIds = bridgeData.map(val => val.chainId).filter((val, index, array) => array.indexOf(val) === index)
-const chainMap: { [chainId: string]: BridgeDataInfo[] } = {}
-
-chainIds.forEach((chainId) => {
-  chainMap[chainId] = bridgeData.filter(val => val.chainId === chainId);
-  const chainInfo = chainMap[chainId][0]
-  chains[chainId] = { id: chainId, icon: chainInfo.logoUrl, name: chainInfo.name, symbol: chainInfo.symbol } as Chain
-})
+import { AvailableChainsInfo, chains, Chain, anyswapInfo} from './interface';
 
 export const DEFAULT_CHAIN_FROM: Chain = chains[0]
 export const DEFAULT_CHAIN_TO: Chain = chains[ChainId.SMARTBCH]
@@ -248,43 +43,24 @@ export default function Bridge() {
 
   const { account: activeAccount, chainId: activeChainId } = useActiveWeb3React()
   const { account, library, activate, chainId } = useWeb3React()
-  const { push } = useRouter()
-
+  const userEthBalance = useETHBalances(account ? [account] : [])?.[account ?? '']
   const bchPrice = useEthPrice()
-
   const transactionUpdater = useTransactionUpdater();
 
-  const currentChainFrom = chainId && chains[chainId] &&
-    { id: chainId, icon: chains[chainId].icon, name: chains[chainId].name }
   const [helpVisible, setHelpVisible] = useState<boolean>(false)
-
   const [chainFrom, setChainFrom] = useState<Chain | null>(DEFAULT_CHAIN_FROM)
   const [chainTo, setChainTo] = useState<Chain | null>(DEFAULT_CHAIN_TO)
   const [shiftAllowed, setShiftAllowed] = useState<boolean>(true)
-  const hopDirection = (chainTo === DEFAULT_CHAIN_TO) ? "in" : "out"
-
   const [methodId, setMethodId] = useState<string | null>(null)
   const [bridgeTransactionHash, setBridgeTransactionHash] = useState<string | null>(null)
   const [destinationAddress, setDestinationAddress] = useState<string | null>("")
-
   const [tokenList, setTokenList] = useState<Currency[] | null>([])
   const [currency0, setCurrency0] = useState<Currency | null>(null)
   const [currencyAmount, setCurrencyAmount] = useState<string | null>('')
   const [tokenToBridge, setTokenToBridge] = useState<AvailableChainsInfo | null>(null)
-  const currencyContract = null // useTokenContract(currency0?.isToken && currency0?.address, true)
-  const anyswapCurrencyContract = null
-  // const anyswapCurrencyContract = useAnyswapTokenContract(
-  //   currency0 && currency0.chainId == ChainId.SMARTBCH && tokenToBridge.other.ContractAddress,
-  //   true
-  // )
-  const [pendingTx, setPendingTx] = useState(false)
   const [showBridgeModal, setShowBridgeModal] = useState(false)
 
-  const selectedCurrencyBalance = useMultichainCurrencyBalance(
-    chainFrom?.id,
-    account ?? undefined,
-    currency0 ?? undefined
-  )
+  const hopDirection = (chainTo === DEFAULT_CHAIN_TO) ? "in" : "out"
 
   useEffect(() => {
     let tokens: Currency[] = Object.keys((anyswapInfo && anyswapInfo[chainFrom.id]) || {})
@@ -387,7 +163,6 @@ export default function Bridge() {
 
   const bridgeButtonClick = () => {
     const address = hopDirection === HopDirection.in ? account : destinationAddress;
-    console.log(methodId)
     const bridgeTransaction = {
       hash: randomId(),
       hopStatus: { destinationAddress: address, direction: hopDirection },
@@ -446,9 +221,12 @@ export default function Bridge() {
   }, [currencyAmount, tokenToBridge])
 
   const insufficientBalance = () => {
-    if (currencyAmount && selectedCurrencyBalance) {
+    if (hopDirection !== HopDirection.out)
+      return false;
+
+    if (currencyAmount) {
       try {
-        const balance = parseFloat(selectedCurrencyBalance.toFixed(currency0.decimals))
+        const balance = parseFloat(userEthBalance.toFixed(currency0.decimals))
         const amount = parseFloat(currencyAmount)
         return amount > balance
       } catch (ex) {
@@ -476,27 +254,13 @@ export default function Bridge() {
     return false
   }
 
-  const getAmountToReceive = () => {
-    if (!tokenToBridge) return 0
-
-    let fee = parseFloat(currencyAmount) * tokenToBridge?.other?.SwapFeeRate
-    if (fee < tokenToBridge?.other?.MinimumSwapFee) {
-      fee = tokenToBridge?.other?.MinimumSwapFee
-    } else if (fee > tokenToBridge?.other?.MaximumSwapFee) {
-      fee = tokenToBridge?.other?.MinimumSwapFee
-    }
-
-    return (parseFloat(currencyAmount) - fee).toFixed(6)
-  }
-
   const buttonDisabled =
     !currency0 ||
     !currencyAmount ||
     currencyAmount == '' ||
     !aboveMin() ||
     !belowMax() ||
-    // insufficientBalance() ||
-    pendingTx
+    insufficientBalance()
 
   const buttonText =
     !shiftAllowed ? `Bridge forbidden (blocked country)` :
@@ -510,74 +274,7 @@ export default function Bridge() {
       ? `Above Maximum Amount`
       : insufficientBalance()
       ? `Insufficient Balance`
-      : pendingTx
-      ? `Confirming Transaction`
       : `Bridge ${currency0?.symbol}`
-
-  // const bridgeToken = async () => {
-  //   const token = tokenToBridge.other
-  //   const depositAddress = currency0.chainId == ChainId.SMARTBCH ? token.ContractAddress : token.DepositAddress
-
-  //   const amountToBridge = ethers.utils.parseUnits(currencyAmount, token.Decimals)
-  //   setPendingTx(true)
-
-  //   try {
-  //     if (currency0.chainId == ChainId.SMARTBCH) {
-  //       if (currency0.isNative) {
-  //       } else if (currency0.isToken) {
-  //         const fn = anyswapCurrencyContract?.interface?.getFunction('Swapout')
-  //         const data = anyswapCurrencyContract.interface.encodeFunctionData(fn, [amountToBridge.toString(), account])
-  //         const tx = await library.getSigner().sendTransaction({
-  //           value: 0x0,
-  //           from: account,
-  //           to: currency0.address,
-  //           data,
-  //         })
-  //         addTransaction(tx, {
-  //           summary: `${i18n._(t`Bridge `)} ${tokenToBridge.symbol}`,
-  //           destChainId: chainTo.id.toString(),
-  //           srcChaindId: chainFrom.id.toString(),
-  //           pairId: tokenToBridge.id,
-  //         })
-  //         push('/bridge/history')
-  //       }
-  //     } else {
-  //       if (currency0.isNative) {
-  //         const tx = await library.getSigner().sendTransaction({
-  //           from: account,
-  //           to: depositAddress,
-  //           value: amountToBridge,
-  //         })
-  //         addTransaction(tx, {
-  //           summary: `${i18n._(t`Bridge `)} ${tokenToBridge.symbol}`,
-  //           destChainId: chainTo.id.toString(),
-  //           srcChaindId: chainFrom.id.toString(),
-  //           pairId: tokenToBridge.id,
-  //         })
-  //         push('/bridge/history')
-  //       } else if (currency0.isToken) {
-  //         const fn = currencyContract?.interface?.getFunction('transfer')
-  //         const data = currencyContract.interface.encodeFunctionData(fn, [depositAddress, amountToBridge.toString()])
-  //         const tx = await library.getSigner().sendTransaction({
-  //           value: 0x0,
-  //           from: account,
-  //           to: currency0.address,
-  //           data,
-  //         })
-  //         addTransaction(tx, {
-  //           summary: `${i18n._(t`Bridge `)} ${tokenToBridge.symbol}`,
-  //           destChainId: chainTo.id.toString(),
-  //           srcChaindId: chainFrom.id.toString(),
-  //           pairId: tokenToBridge.id,
-  //         })
-  //         push('/bridge/history')
-  //       }
-  //     }
-  //   } catch (ex) {
-  //   } finally {
-  //     setPendingTx(false)
-  //   }
-  // }
 
   const anyswapChains = Object.keys(anyswapInfo).map(val => parseInt(val)) //[ChainId.SMARTBCH]
   let availableChains = Object.keys(anyswapInfo || {})
@@ -593,41 +290,14 @@ export default function Bridge() {
         isOpen={showBridgeModal}
         hash={bridgeTransactionHash}
         onDismiss={() => setShowBridgeModal(false)} />)}
-      {/* <Modal isOpen={showBridgeModal} onDismiss={() => setShowBridgeModal(false)}>
-        <div className="space-y-4">
-          <ModalHeader title={i18n._(t`Bridge ${currency0?.symbol}`)} onClose={() => setShowBridgeModal(false)} />
-          <Typography variant="sm" className="font-medium">
-            {i18n._(t`You are sending ${formatNumber(currencyAmount)} ${currency0?.symbol} from ${chainFrom?.name}`)}
-          </Typography>
-          <Typography variant="sm" className="font-medium">
-            {i18n._(t`You will receive ${formatNumber(getAmountToReceive())} ${currency0?.symbol} on ${chainTo?.name}`)}
-          </Typography>
-
-          <Button color="gradient" size="lg" disabled={pendingTx} onClick={() => bridgeToken()}>
-            <Typography variant="lg">
-              {pendingTx ? (
-                <div className={'p-2'}>
-                  <AutoRow gap="6px" justify="center">
-                    {buttonText} <Loader stroke="white" />
-                  </AutoRow>
-                </div>
-              ) : (
-                i18n._(t`Bridge ${currency0?.symbol}`)
-              )}
-            </Typography>
-          </Button>
-        </div>
-      </Modal> */}
 
       <Head>
         <title>{i18n._(t`Bridge`)} | MISTswap</title>
         <meta key="description" name="description" content="Bridge" />
       </Head>
 
-      {/* <SolarbeamLogo /> */}
-
-      <Container maxWidth="2xl" className="space-y-6">
-        <DoubleGlowShadow /*opacity="0.6"*/>
+      <Container maxWidth="2xl" className="mt-5 space-y-6">
+        <DoubleGlowShadow>
           <div className="p-4 space-y-4 rounded bg-dark-900" style={{ zIndex: 1 }}>
             <div className="flex items-center justify-center mb-4 space-x-3">
               <div className="grid grid-cols-2 rounded p-3px bg-dark-800 h-[46px]">
@@ -706,7 +376,15 @@ export default function Bridge() {
                 switchOnSelect={true}
               />
               <button className={'sm:m-6'}>
-                <ArrowRight size="32" />
+                <ArrowRight size="32" onClick={() => {
+                  if (chainTo.id === ChainId.SMARTBCH) {
+                    setChainTo(chains[0])
+                    setChainFrom(chains[ChainId.SMARTBCH])
+                  } else {
+                    setChainTo(chains[ChainId.SMARTBCH])
+                    setChainFrom(chains[0])
+                  }
+                }} />
               </button>
               <ChainSelect
                 chains={chains}
@@ -757,15 +435,7 @@ export default function Bridge() {
                   size="lg"
                   disabled={buttonDisabled}
                 >
-                  {pendingTx ? (
-                    <div className={'p-2'}>
-                      <AutoRow gap="6px" justify="center">
-                        {buttonText} <Loader stroke="white" />
-                      </AutoRow>
-                    </div>
-                  ) : (
-                    buttonText
-                  )}
+                  {buttonText}
                 </Button>
               )}
             </BottomGrouping>
